requirements.txt
flet
opencv-python
pytesseract
numpy
mysql-connector-python
pandas
pyinstaller